requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn


accelerate

bitsandbytes

langchain_community

grandalf

tiktoken

pypdf

bs4

langchain_experimental


chromadb

faiss-cpu


youtube-transcript-api


flask

duckduckgo-search

langchain-google-community


pydantic

gTTS

# For speech to text
SpeechRecognition

PyAudio

# For text to speech
gTTS

playsound


gTTS
pydub

pygame


websockets

fastapi 
uvicorn 

markdown


langchain_groq
playwright 

beautifulsoup4



flask
python-dotenv 
youtube-transcript-api 
langchain-google-genai 
langchain 
faiss-cpu